corrected issue #410 Add to refactor a bit different visitors
